Move old/reference Neff calculation into its own file
